requirements.txt
fastapi
pymysql
uvicorn
python-dotenv
